change actions, add state
