+++ src/store/actions/checkOauthWindow.jsx
