+++ src/store/actions/types.jsx
@@ -1,4 +1,3 @@
 export const OPEN_WINDOW = 'OPEN_WINDOW';
 export const CLOSE_WINDOW = 'CLOSE_WINDOW';
 export const CHECK_WINDOW = 'CHECK_WINDOW';
-export const GENERATE_STATE = 'GENERATE_STATE';
