+++ src/store/index.jsx
@@ -2,7 +2,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import rootReducer from '../reducers/index';
+import rootReducer from './reducers/index';
 
 // const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION__ || compose;
 const store = createStore(
